chore: replace babel with next
