+++ packages/babel-plugin/package.json
@@ -19,17 +19,16 @@
     "src"
   ],
   "dependencies": {
-    "@babel/core": "^7.9.0",
+    "@babel/core": "^7.10.2",
     "@babel/generator": "^7.10.2",
-    "@babel/helper-plugin-utils": "^7.10.1",
-    "@babel/plugin-syntax-jsx": "^7.8.3",
-    "@babel/plugin-transform-flow-strip-types": "^7.8.3",
-    "@babel/template": "^7.10.1",
-    "@babel/traverse": "^7.10.1",
+    "@babel/helper-plugin-utils": "^7.10.2",
+    "@babel/plugin-syntax-jsx": "^7.10.2",
+    "@babel/plugin-transform-flow-strip-types": "^7.10.2",
+    "@babel/template": "^7.10.2",
+    "@babel/traverse": "^7.10.2",
     "@babel/types": "^7.10.2",
     "@compiled/ts-transform-css-in-js": "0.4.7",
-    "@emotion/is-prop-valid": "^0.8.6",
-    "typescript": "^3.7.3"
+    "@emotion/is-prop-valid": "^0.8.6"
   },
   "devDependencies": {
     "@types/babel__core": "^7.1.7"
